Database güncellemeleri ve ödeme hataları ayarlandı

+++ API/Controllers/OrdersController.cs
@@ -1,4 +1,7 @@
-
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Threading.Tasks;
 using API.Data;
 using API.DTO;
 using API.Entity;
@@ -18,12 +21,11 @@ namespace API.Controllers
     public class OrdersController : ControllerBase
     {
         private readonly DataContext _context;
-        private readonly IConfiguration _configuration;
-        public OrdersController(DataContext context, IConfiguration configuration)
+        private readonly IConfiguration _config;
+        public OrdersController(DataContext context, IConfiguration config)
         {
             _context = context;
-            _configuration = configuration;
-
+            _config = config;
         }
 
         [HttpGet]
@@ -57,18 +59,18 @@ namespace API.Controllers
 
             if (cart == null) return BadRequest(new ProblemDetails { Title = "Problem getting cart" });
 
-            var items = new List<DTO.OrderItem>();
+            var items = new List<Entity.OrderItem>();
 
             foreach (var item in cart.CartItems)
             {
                 var product = await _context.Products.FindAsync(item.ProductId);
 
-                var orderItem = new DTO.OrderItem
+                var orderItem = new Entity.OrderItem
                 {
                     ProductId = product!.Id,
                     ProductName = product.Name!,
                     ProductImage = product.ImageUrl!,
-                    Price = (decimal)product.Price!,
+                    Price =  (decimal)product.Price,
                     Quantity = item.Quantity
                 };
 
@@ -82,7 +84,7 @@ namespace API.Controllers
             var order = new Order
             {
                 OrderItems = items,
-                CustomerId = User.Identity!.Name!,
+                CustomerId = User.Identity!.Name,
                 FirstName = orderDTO.FirstName,
                 LastName = orderDTO.LastName,
                 Phone = orderDTO.Phone,
@@ -91,8 +93,17 @@ namespace API.Controllers
                 SubTotal = subTotal,
                 DeliveryFee = deliveryFee
             };
-            //Payment
-             await ProcessPayment(orderDTO,cart);
+
+            var paymentResult = await ProcessPayment(orderDTO, cart);
+
+            if (paymentResult.Status == "failure")
+            {
+                return BadRequest(new ProblemDetails { Title = paymentResult.ErrorMessage });
+            }
+
+            order.ConversationId = paymentResult.ConversationId;
+            order.BasketId = paymentResult.BasketId;
+
             _context.Orders.Add(order);
             _context.Carts.Remove(cart);
 
@@ -103,75 +114,77 @@ namespace API.Controllers
 
             return BadRequest(new ProblemDetails { Title = "Problem getting order" });
         }
-    private async Task<Payment> ProcessPayment(CreateOrderDTO model,Cart cart)
-    {
-        Options options = new Options();
-        options.ApiKey = _configuration["PaymentAPI:APIKey"];
-        options.SecretKey = _configuration["PaymentAPI:SecretKey"];
-        options.BaseUrl = "https://sandbox-api.iyzipay.com";
-                
-        CreatePaymentRequest request = new CreatePaymentRequest();
-        request.Locale = Locale.TR.ToString();
-        request.ConversationId = Guid.NewGuid().ToString();
-        request.Price = cart.CalculateTotal().ToString();
-        request.PaidPrice = cart.CalculateTotal().ToString();
-        request.Currency = Currency.TRY.ToString();
-        request.Installment = 1;
-        request.BasketId =  Guid.NewGuid().ToString();
-        request.PaymentChannel = PaymentChannel.WEB.ToString();
-        request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
 
-        PaymentCard paymentCard = new PaymentCard();
-        paymentCard.CardHolderName = model.CardName;
-        paymentCard.CardNumber = model.CardNumber;
-        paymentCard.ExpireMonth = model.CardExpireMonth;
-        paymentCard.ExpireYear = model.CardExpireYear;
-        paymentCard.Cvc = model.CardCvc;
-        paymentCard.RegisterCard = 0;
-        request.PaymentCard = paymentCard;
-
-        Buyer buyer = new Buyer();
-        buyer.Id = "BY789";
-        buyer.Name = model.FirstName;
-        buyer.Surname = model.LastName;
-        buyer.GsmNumber = model.Phone;
-        buyer.Email = "email@email.com";
-        buyer.IdentityNumber = "74300864791";
-        buyer.LastLoginDate = "2015-10-05 12:43:35";
-        buyer.RegistrationDate = "2013-04-21 15:12:09";
-        buyer.RegistrationAddress = model.Address;
-        buyer.Ip = "85.34.78.112";
-        buyer.City = model.City;
-        buyer.Country = "Turkiye";
-        buyer.ZipCode = "34732";
-        request.Buyer = buyer;
-
-        Address shippingAddress = new Address();
-        shippingAddress.ContactName = model.FirstName + " " + model.LastName;
-        shippingAddress.City = model.City;
-        shippingAddress.Country = "Türkiye";
-        shippingAddress.Description = model.Address;
-        shippingAddress.ZipCode = "34742";
-        request.ShippingAddress = shippingAddress;
-        request.BillingAddress = shippingAddress;
-
-
-        List<BasketItem> basketItems = new List<BasketItem>();
-        foreach (var item in cart.CartItems)
+        private async Task<Payment> ProcessPayment(CreateOrderDTO model, Cart cart)
         {
-            BasketItem basketItem = new BasketItem();
-            basketItem.Id = item.ProductId.ToString();
-            basketItem.Name = item.Product.Name;
-            basketItem.Category1 = "Telefon";
-            basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
-            basketItem.Price = ((double)item.Product.Price! * item.Quantity).ToString();
-            basketItems.Add(basketItem );
+            Options options = new Options();
+            options.ApiKey = _config["PaymentAPI:APIKey"];
+            options.SecretKey = _config["PaymentAPI:SecretKey"];
 
+            options.BaseUrl = "https://sandbox-api.iyzipay.com";
+
+            CreatePaymentRequest request = new CreatePaymentRequest();
+            request.Locale = Locale.TR.ToString();
+            request.ConversationId = Guid.NewGuid().ToString();
+            request.Price = cart.CalculateTotal().ToString();
+            request.PaidPrice = cart.CalculateTotal().ToString();
+            request.Currency = Currency.TRY.ToString();
+            request.Installment = 1;
+            request.BasketId = cart.CartId.ToString();
+            request.PaymentChannel = PaymentChannel.WEB.ToString();
+            request.PaymentGroup = PaymentGroup.PRODUCT.ToString();
+
+            PaymentCard paymentCard = new PaymentCard();
+            paymentCard.CardHolderName = model.CardName;
+            paymentCard.CardNumber = model.CardNumber;
+            paymentCard.ExpireMonth = model.CardExpireMonth;
+            paymentCard.ExpireYear = model.CardExpireYear;
+            paymentCard.Cvc = model.CardCvc;
+            paymentCard.RegisterCard = 0;
+            request.PaymentCard = paymentCard;
+
+            Buyer buyer = new Buyer();
+            buyer.Id = "BY789";
+            buyer.Name = model.FirstName;
+            buyer.Surname = model.LastName;
+            buyer.GsmNumber = model.Phone;
+            buyer.Email = "email@email.com";
+            buyer.IdentityNumber = "74300864791";
+            buyer.LastLoginDate = "2015-10-05 12:43:35";
+            buyer.RegistrationDate = "2013-04-21 15:12:09";
+            buyer.RegistrationAddress = model.Address;
+            buyer.Ip = "85.34.78.112";
+            buyer.City = model.City;
+            buyer.Country = "Türkiye";
+            buyer.ZipCode = "34732";
+            request.Buyer = buyer;
+
+            Address shippingAddress = new Address();
+            shippingAddress.ContactName = model.FirstName + " " + model.LastName;
+            shippingAddress.City = model.City;
+            shippingAddress.Country = "Türkiye";
+            shippingAddress.Description = model.Address;
+            shippingAddress.ZipCode = "34742";
+
+            request.ShippingAddress = shippingAddress;
+            request.BillingAddress = shippingAddress;
+
+            List<BasketItem> basketItems = new List<BasketItem>();
+
+            foreach (var item in cart.CartItems)
+            {
+                BasketItem basketItem = new BasketItem();
+                basketItem.Id = item.ProductId.ToString();
+                basketItem.Name = item.Product.Name;
+                basketItem.Category1 = "Saat";
+                basketItem.ItemType = BasketItemType.PHYSICAL.ToString();
+                basketItem.Price = ((double)item.Product.Price * item.Quantity).ToString();
+                basketItems.Add(basketItem);
+            }
+
+            request.BasketItems = basketItems;
+
+            return await Payment.Create(request, options);
         }
-        request.BasketItems = basketItems;
-       
-
-        return await Payment.Create(request, options);
-}
-}
+    }
 }
